removed prefetch query and created form component to handle undefined default form values
